AutoComplete clean-up
