+++ packages/desktop-client/src/components/util/GenericInput.jsx
@@ -8,7 +8,7 @@ import { AccountAutocomplete } from '../autocomplete/AccountAutocomplete';
 import { Autocomplete } from '../autocomplete/Autocomplete';
 import { CategoryAutocomplete } from '../autocomplete/CategoryAutocomplete';
 import { PayeeAutocomplete } from '../autocomplete/PayeeAutocomplete';
-import { SavedFilterAutocomplete } from '../autocomplete/SavedFilterAutocomplete';
+import { FilterAutocomplete } from '../autocomplete/FilterAutocomplete';
 import { Input } from '../common/Input';
 import { View } from '../common/View';
 import { Checkbox } from '../forms';
@@ -99,7 +99,7 @@ export function GenericInput({
       switch (field) {
         case 'saved':
           content = (
-            <SavedFilterAutocomplete
+            <FilterAutocomplete
               saved={saved}
               value={value}
               multi={multi}
